refactor: refactored AudioVisualizer example
