+++ apps/common-app/src/examples/AudioVisualizer/AudioVisualizer.tsx
@@ -7,6 +7,7 @@ import {
   AudioBufferSourceNode,
 } from 'react-native-audio-api';
 import { ActivityIndicator, View, StyleSheet } from 'react-native';
+import { useSharedValue } from 'react-native-reanimated';
 
 import FreqTimeChart from './FreqTimeChart';
 import { Container, Button } from '../../components';
@@ -21,11 +22,8 @@ const AudioVisualizer: React.FC = () => {
   const [isPlaying, setIsPlaying] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
 
-  const [times, setTimes] = useState<number[]>(
-    new Array(FFT_SIZE / 2).fill(127)
-  );
-
-  const [freqs, setFreqs] = useState<number[]>(new Array(FFT_SIZE / 2).fill(0));
+  const times = useSharedValue<number[]>(new Array(FFT_SIZE / 2).fill(127));
+  const freqs = useSharedValue<number[]>(new Array(FFT_SIZE / 2).fill(0));
 
   const audioContextRef = useRef<AudioContext | null>(null);
   const analyserRef = useRef<AnalyserNode | null>(null);
@@ -61,11 +59,11 @@ const AudioVisualizer: React.FC = () => {
 
     const timesArray = new Array(bufferLength);
     analyserRef.current.getByteTimeDomainData(timesArray);
-    setTimes(timesArray);
+    times.value = timesArray;
 
     const freqsArray = new Array(bufferLength);
     analyserRef.current.getByteFrequencyData(freqsArray);
-    setFreqs(freqsArray);
+    freqs.value = freqsArray;
 
     requestAnimationFrame(draw);
   };
