+++ apps/common-app/src/examples/AudioVisualizer/FreqTimeChart.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
+import { SharedValue } from 'react-native-reanimated';
 
 import { withCanvas } from './Canvas';
 import Charts from './Charts';
 
 interface FreqTimeChartProps {
-  timeData: number[];
-  frequencyData: number[];
+  timeData: SharedValue<number[]>;
+  frequencyData: SharedValue<number[]>;
   frequencyBinCount: number;
 }
 
@@ -14,11 +15,8 @@ const FreqTimeChart: React.FC<FreqTimeChartProps> = (props) => {
 
   return (
     <>
-      <Charts.TimeChartLine
-        data={timeData}
-        frequencyBinCount={frequencyBinCount}
-      />
-      <Charts.FrequencyChartLine
+      <Charts.TimeChart data={timeData} frequencyBinCount={frequencyBinCount} />
+      <Charts.FrequencyChart
         data={frequencyData}
         frequencyBinCount={frequencyBinCount}
       />
