+++ apps/common-app/src/examples/DrumMachine/PatternShape.tsx
@@ -1,10 +1,10 @@
 import React, { useMemo } from 'react';
 import { Circle, Paint, Path, vec, Skia } from '@shopify/react-native-skia';
 
-import { getAngle, getPointCX, getPointCY } from './utils';
+import { getAngle, getPointCX, getPointCY } from '../../utils/skiUtils';
 import instruments from './instruments';
 import type { Pattern } from '../../types';
-import { buttonRadius, cPoint } from './constants';
+import { buttonRadius, cPoint, numBeats } from './constants';
 
 interface PatternShapeProps {
   pattern: Pattern;
@@ -29,7 +29,7 @@ const PatternShape: React.FC<PatternShapeProps> = (props) => {
           return null;
         }
 
-        const angle = getAngle(index);
+        const angle = getAngle(index, numBeats);
         const x = getPointCX(angle, instrument.radius, cPoint.x);
         const y = getPointCY(angle, instrument.radius, cPoint.y);
