+++ apps/common-app/src/utils/skiUtils.ts
@@ -1,9 +1,5 @@
-import { numBeats } from './constants';
-
-export function getAngle(stepIdx: number) {
-  'worklet';
-
-  return (stepIdx / numBeats) * Math.PI * 2 - Math.PI / 2;
+export function getAngle(stepIdx: number, maxSteps: number) {
+  return (stepIdx / maxSteps) * Math.PI * 2 - Math.PI / 2;
 }
 
 export function getPointCX(angle: number, radius: number, outerCX: number) {
